ReactlessBuster2/requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
gspread>=5.12.0
oauth2client>=4.1.3
aiohttp>=3.9.0
python-dotenv>=1.0.0
